refactor(domain): move incident window, alert extraction, and diagnosis helpers
Relocate pure investigation modules into app/core/domain and keep
orchestration nodes as thin LLM/I/O glue. Add operator-hint scoring to
domain correlation and unit tests for new extractions.

Fixes #3018
