+++ app/tools/GitDeployTimelineTool/__init__.py
@@ -22,7 +22,7 @@ from __future__ import annotations
 from datetime import UTC, datetime, timedelta
 from typing import Any
 
-from app.incident_window import IncidentWindow
+from app.core.domain.types.incident_window import IncidentWindow
 from app.integrations.github_mcp import call_github_mcp_tool
 from app.tools.tool_decorator import tool
 from app.tools.utils.github_helpers import (
