+++ tests/agent/test_result.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 
-from app.core.orchestration.node.diagnose.node import (
-    _build_diagnosis_schema,
-    _extract_last_assistant_text,
-    _taxonomy_categories_for_alert_source,
+from app.core.domain.state.diagnosis import (
+    build_diagnosis_schema,
+    extract_last_assistant_text,
+    taxonomy_categories_for_alert_source,
 )
 
 
@@ -32,24 +32,22 @@ def test_extract_last_assistant_text_handles_anthropic_content_blocks() -> None:
         },
     ]
 
-    assert _extract_last_assistant_text(messages) == (
-        "## Diagnosis\n Root cause: missing telemetry"
-    )
+    assert extract_last_assistant_text(messages) == ("## Diagnosis\n Root cause: missing telemetry")
 
 
 def test_non_hermes_taxonomy_excludes_hermes_categories() -> None:
-    categories = _taxonomy_categories_for_alert_source("postgresql")
+    categories = taxonomy_categories_for_alert_source("postgresql")
     assert "agent_hang" not in categories
     assert "ghost_session" not in categories
 
 
 def test_hermes_taxonomy_is_scoped_to_hermes_categories() -> None:
-    categories = _taxonomy_categories_for_alert_source("hermes")
+    categories = taxonomy_categories_for_alert_source("hermes")
     assert "agent_hang" in categories
     assert "ghost_session" in categories
     assert "connection_exhaustion" not in categories
 
-    schema = _build_diagnosis_schema(categories)
+    schema = build_diagnosis_schema(categories)
     description = str(schema.model_fields["root_cause_category"].description)
     assert "agent_hang" in description
     assert "connection_exhaustion" not in description
